requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
PyAudio==0.2.14
wikipedia==1.4.0
python-dotenv==1.0.0
psutil==5.9.5
requests==2.31.0
PyAutoGUI==0.9.54
datetime==5.1
pyinstaller==6.1.0
pillow==10.2.0
python-weather==1.1.1
googlesearch-python==1.2.3
newsapi-python==0.2.7
pyperclip==1.8.2
screen-brightness-control==0.22.1
pyjokes==0.6.0
wolframalpha==5.0.0
playsound==1.3.0
pynput==1.7.6
